UI time-series management : renaming TS management renderer files with more proper names
